Address review feedback on artifact post-processing
- Parse the merge-host timeout override as long rather than int. A value
  above int.MaxValue chosen to mean 'a very long time' overflowed the parse
  and silently fell back to the 15-minute default, the opposite of what the
  caller asked for, while a merely large value correctly became unbounded.

- Remove TESTINGPLATFORM_DOTNETTEST_EXECUTIONID from the child environment
  in the end-to-end tests instead of setting it to an empty string. The test
  framework already exposes EnvironmentToRemove, which genuinely unsets the
  variable, so the test no longer depends on the platform treating empty and
  unset alike, and it now does what its comment says.

Copilot-Session: a41f8680-43fb-4285-b81e-ed5a2788fbd2

+++ src/Cli/dotnet/Commands/Test/MTP/TestApplication.cs
@@ -415,7 +415,11 @@ internal sealed class TestApplication(
 
     internal static TimeSpan ParseArtifactPostProcessingTimeout(string? configuredTimeout)
     {
-        if (!int.TryParse(configuredTimeout, NumberStyles.Integer, CultureInfo.InvariantCulture, out int seconds) || seconds < 0)
+        // Parsed as long rather than int so that a value chosen to mean 'a very long time' lands in
+        // the 'effectively never' branch below instead of overflowing the parse and silently falling
+        // back to the default. Anything above long.MaxValue seconds is not a duration anyone means,
+        // so it keeps the default along with the other unusable values.
+        if (!long.TryParse(configuredTimeout, NumberStyles.Integer, CultureInfo.InvariantCulture, out long seconds) || seconds < 0)
         {
             return DefaultArtifactPostProcessingTimeout;
         }
